MAUI - [x] Code refactoring and suggested changes on the PR for Classic component BarcodeSDK MAUI.
