+++ ScanbotBarcodeSDKExample/BarcodeSDK.MAUI.Example/BarcodeSDK.MAUI.Example.csproj
@@ -68,6 +68,7 @@
         <Folder Include="Models\" />
         <Folder Include="ClassicComponent\" />
         <Folder Include="Platforms\Android\Resources\layout\" />
+        <Folder Include="Platforms\iOS\Utils\" />
     </ItemGroup>
     <ItemGroup Condition="$(TargetFramework.StartsWith('net7.0-android')) != true">
         <Compile Remove="**\**\*.Android.cs" />
@@ -84,5 +85,8 @@
     </ItemGroup>
     <ItemGroup>
       <AndroidResource Include="Platforms\Android\Resources\layout\barcode_camera_view.xml" />
+    </ItemGroup>
+    <ItemGroup>
+      <None Remove="Platforms\iOS\Utils\" />
     </ItemGroup>
 </Project>
